refactor(sanity): extract sanity resource related functions
- Created new `resource.ts` module to centralize Sanity resource management functions
- Moved `writeTagsToSanityResource` and `syncSanityResourceInstructor` from `course.ts`
- Added utility functions `getSanityResourceForPost` and `getSoftwareLibraryReferencesFromTags`
- Introduced `createVersionedSoftwareLibraries` to handle software library references
- Updated Sanity resource tag handling to use a more structured approach
- Improved code organization and separation of concerns
