+++ apps/egghead/src/lib/sanity/course.ts
@@ -9,6 +9,7 @@ import { eq } from 'drizzle-orm'
 
 import { getSanityCollaborator } from './collaborator'
 import { getSanityLessonForEggheadLessonId } from './lesson'
+import { writeTagsToSanityResource } from './resource'
 import { getSanitySoftwareLibrary } from './softwarelibrary'
 import type {
 	PositionInputItem,
@@ -239,140 +240,4 @@ export async function updateSanityCourseMetadata({
 			image,
 		})
 		.commit()
-}
-
-/**
- * Writes tags to a Sanity resource (lesson or course)
- * @param postId - The ID of the post
- * @returns The updated Sanity resource document
- */
-export async function writeTagsToSanityResource(postId: string) {
-	const post = await getPost(postId)
-
-	if (!post) {
-		throw new Error(`Post with id ${postId} not found.`)
-	}
-
-	let sanityResource
-	if (post.fields.postType === 'lesson') {
-		sanityResource = await getSanityLessonForEggheadLessonId(
-			post.fields.eggheadLessonId,
-		)
-	} else if (post.fields.postType === 'course') {
-		sanityResource = await getSanityCourseForEggheadCourseId(
-			post.fields.eggheadPlaylistId,
-		)
-	}
-
-	if (!sanityResource) {
-		throw new Error(`Sanity resource for post with id ${postId} not found.`)
-	}
-
-	const softwareLibraries = post?.tags
-		? await Promise.all(
-				post?.tags?.map(async ({ tag }) => {
-					return await getSanitySoftwareLibrary(tag.fields.slug)
-				}),
-			)
-		: []
-
-	const softwareLibraryArrayReferences = softwareLibraries.map((library) => {
-		return createSanityArrayElementReference(library?._id || '')
-	})
-
-	return await sanityWriteClient
-		.patch(sanityResource._id)
-		.set({
-			softwareLibraryArrayReferences,
-		})
-		.commit()
-}
-
-/**
- * Syncs a Sanity resource instructor
- * @param postId - The ID of the post
- * @param userId - The ID of the user
- * @returns The updated Sanity resource document
- */
-export async function syncSanityResourceInstructor(
-	postId: string,
-	userId: string,
-) {
-	const { ability, session } = await getServerAuthSession()
-
-	if (!session?.user?.id || ability.cannot('manage', 'all')) {
-		throw new Error('Unauthorized')
-	}
-
-	const post = await getPost(postId)
-
-	if (!post) {
-		throw new Error(`Post with id ${postId} not found.`)
-	}
-
-	// Extract instructor ID using a more flexible approach
-	let instructorId: string | number | null = null
-
-	// Use a type guard to safely access properties
-	const fields = post.fields as Record<string, any>
-
-	if (
-		fields.instructor &&
-		typeof fields.instructor === 'object' &&
-		'id' in fields.instructor
-	) {
-		instructorId = fields.instructor.id
-	} else if ('instructorId' in fields) {
-		instructorId = fields.instructorId
-	}
-
-	if (!instructorId) {
-		throw new Error(`Instructor ID not found for post ${postId}.`)
-	}
-
-	const sanityInstructorReference = await getSanityCollaborator(
-		parseInt(instructorId.toString()),
-	)
-
-	if (!sanityInstructorReference) {
-		throw new Error(`Sanity instructor with id ${instructorId} not found.`)
-	}
-
-	switch (post.fields.postType) {
-		case 'lesson':
-			const sanityLesson = await getSanityLessonForEggheadLessonId(
-				post.fields.eggheadLessonId,
-			)
-
-			if (!sanityLesson || !sanityLesson._id) {
-				throw new Error(
-					`Sanity lesson with id ${post.fields.eggheadLessonId} not found.`,
-				)
-			}
-
-			return await sanityWriteClient
-				.patch(sanityLesson?._id)
-				.set({
-					collaborators: [sanityInstructorReference],
-				})
-				.commit()
-
-		case 'course':
-			const sanityCourse = await getSanityCourseForEggheadCourseId(
-				post.fields.eggheadPlaylistId,
-			)
-
-			if (!sanityCourse || !sanityCourse._id) {
-				throw new Error(
-					`Sanity course with id ${post.fields.eggheadPlaylistId} not found.`,
-				)
-			}
-
-			return await sanityWriteClient
-				.patch(sanityCourse?._id)
-				.set({
-					collaborators: [sanityInstructorReference],
-				})
-				.commit()
-	}
 }
